requirements.txt
Django>=5.1.1
psycopg2>=2.9.9
PyJWT>=2.9.0
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.1